feat(web/app/dashboard/navigation): sidebar according to roles
